refactor(uploader): reorganize into problems/ and categories/; move upload.py under problems/; update scripts + workflows
